move XML files to XML folder
