+++ IslandMUD/IslandMUD.vcxproj
@@ -103,7 +103,7 @@
     <ClCompile Include="NPC\npc_unaffiliated.cpp" />
     <ClCompile Include="parse.cpp" />
     <ClCompile Include="player_character.cpp" />
-    <ClCompile Include="pugixml.cpp" />
+    <ClCompile Include="XML\pugixml.cpp" />
     <ClCompile Include="recipes.cpp" />
     <ClCompile Include="room.cpp" />
     <ClCompile Include="room_side.cpp" />
@@ -137,8 +137,8 @@
     <ClInclude Include="NPC\npc_unaffiliated.h" />
     <ClInclude Include="parse.h" />
     <ClInclude Include="player_character.h" />
-    <ClInclude Include="pugiconfig.hpp" />
-    <ClInclude Include="pugixml.hpp" />
+    <ClInclude Include="XML\pugiconfig.hpp" />
+    <ClInclude Include="XML\pugixml.hpp" />
     <ClInclude Include="recipe.h" />
     <ClInclude Include="recipes.h" />
     <ClInclude Include="resource.h" />
