+++ IslandMUD/XML/pugiconfig.hpp
