+++ IslandMUD/XML/pugixml.cpp
