+++ IslandMUD/XML/pugixml.hpp
